+++ app/Http/Controllers/IndividualWorkController.php
@@ -5,13 +5,16 @@ namespace App\Http\Controllers;
 use App\Http\Requests\IndividualWorkRequest;
 use App\Models\Group;
 use App\Models\IndividualWork;
+use App\Models\Student;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Inertia\Inertia;
 
 class IndividualWorkController extends Controller
 {
     public function create(Request $request, Group $group, string $studentNumber)
     {
+        Gate::authorize('individualWorks', Student::class);
         $type = $request->get('type', 'relative');
         $student = $group->findStudentByNumber($studentNumber);
         return Inertia::render('individual-work/CreatePage', compact('group', 'studentNumber', 'student', 'type'));
@@ -19,6 +22,7 @@ class IndividualWorkController extends Controller
 
     public function store(IndividualWorkRequest $request, Group $group, string $studentNumber)
     {
+        Gate::authorize('individualWorks', Student::class);
         $student = $group->findStudentByNumber($studentNumber);
         $student->individualWork()->create($request->validated());
         return to_route('groups.students.show', ['group' => $group->id, 'student' => $studentNumber]);
@@ -26,6 +30,7 @@ class IndividualWorkController extends Controller
 
     public function edit(Group $group, string $studentNumber, IndividualWork $individualWork)
     {
+        Gate::authorize('individualWorks', Student::class);
         $student = $group->findStudentByNumber($studentNumber);
         return Inertia::render(
             'individual-work/EditPage',
@@ -39,12 +44,14 @@ class IndividualWorkController extends Controller
         string $studentNumber,
         IndividualWork $individualWork
     ) {
+        Gate::authorize('individualWorks', Student::class);
         $individualWork->update($request->validated());
         return to_route('groups.students.show', ['group' => $group->id, 'student' => $studentNumber]);
     }
 
     public function destroy(Group $group, string $studentNumber, IndividualWork $individualWork)
     {
+        Gate::authorize('individualWorks', Student::class);
         $individualWork->delete();
         return to_route('groups.students.show', ['group' => $group->id, 'student' => $studentNumber]);
     }
